Create `replyToOnlyInGroup` method in the `MessageBuilder`.
It's a common behavior of the bot to reply (send the messages as a reply) to the messages in the groups, but in private chats it's preferred to do that less.

So this method replies to the message only if it was in a group (it can happen in channels causing an Api exception too).